+++ src/main/java/lotto/domain/lotto/LottoLine.java
@@ -1,9 +1,10 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+import lotto.domain.Rank;
 
 public class LottoLine {
     private final Set<LottoNumber> value;
