+++ src/main/java/lotto/domain/lotto/LottoTicket.java
@@ -1,7 +1,10 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import java.util.ArrayList;
 import java.util.List;
+import lotto.domain.Rank;
+import lotto.domain.lotto.LottoLine;
+import lotto.domain.lotto.LottoNumber;
 
 public class LottoTicket {
