+++ src/main/java/lotto/domain/lotto/LottoTicketFactory.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import java.util.ArrayList;
