+++ src/test/java/domain/RankTest.java
@@ -1,3 +1,5 @@
+package domain;
+
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import lotto.domain.Rank;
