+++ src/test/java/domain/lotto/LottoNumberTest.java
@@ -1,8 +1,9 @@
+package domain.lotto;
+
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import lotto.domain.LottoNumber;
-import lotto.domain.Money;
+import lotto.domain.lotto.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
